+++ src/main/java/keystrokesmod/module/impl/combat/HitBox.java
@@ -94,14 +94,14 @@ public class HitBox extends Module {
             Vec3 vec5 = vec3.addVector(vec4.xCoord * d0, vec4.yCoord * d0, vec4.zCoord * d0);
             Vec3 vec6 = null;
             float f1 = 1.0F;
-            List list = mc.theWorld.getEntitiesWithinAABBExcludingEntity(mc.getRenderViewEntity(), mc.getRenderViewEntity().getEntityBoundingBox().addCoord(vec4.xCoord * d0, vec4.yCoord * d0, vec4.zCoord * d0).expand((double) f1, (double) f1, (double) f1));
+            List list = mc.theWorld.getEntitiesWithinAABBExcludingEntity(mc.getRenderViewEntity(), mc.getRenderViewEntity().getEntityBoundingBox().addCoord(vec4.xCoord * d0, vec4.yCoord * d0, vec4.zCoord * d0).expand(f1, f1, f1));
             double d3 = d2;
 
             for (Object o : list) {
                 Entity entity = (Entity) o;
                 if (entity.canBeCollidedWith()) {
                     float ex = (float) ((double) entity.getCollisionBorderSize() * getExpand(entity));
-                    AxisAlignedBB ax = entity.getEntityBoundingBox().expand((double) ex, (double) ex, (double) ex);
+                    AxisAlignedBB ax = entity.getEntityBoundingBox().expand(ex, ex, ex);
                     MovingObjectPosition mop = ax.calculateIntercept(vec3, vec5);
                     if (ax.isVecInside(vec3)) {
                         if (0.0D < d3 || d3 == 0.0D) {
@@ -143,7 +143,7 @@ public class HitBox extends Module {
             double y = e.lastTickPosY + (e.posY - e.lastTickPosY) * (double) Utils.getTimer().renderPartialTicks - mc.getRenderManager().viewerPosY;
             double z = e.lastTickPosZ + (e.posZ - e.lastTickPosZ) * (double) Utils.getTimer().renderPartialTicks - mc.getRenderManager().viewerPosZ;
             float ex = (float) ((double) e.getCollisionBorderSize() * multiplier.getInput());
-            AxisAlignedBB bbox = e.getEntityBoundingBox().expand((double) ex, (double) ex, (double) ex);
+            AxisAlignedBB bbox = e.getEntityBoundingBox().expand(ex, ex, ex);
             AxisAlignedBB axis = new AxisAlignedBB(bbox.minX - e.posX + x, bbox.minY - e.posY + y, bbox.minZ - e.posZ + z, bbox.maxX - e.posX + x, bbox.maxY - e.posY + y, bbox.maxZ - e.posZ + z);
             GL11.glBlendFunc(770, 771);
             GL11.glEnable(3042);
@@ -151,7 +151,7 @@ public class HitBox extends Module {
             GL11.glDisable(2929);
             GL11.glDepthMask(false);
             GL11.glLineWidth(2.0F);
-            GL11.glColor3d((double) c.getRed(), (double) c.getGreen(), (double) c.getBlue());
+            GL11.glColor3d(c.getRed(), c.getGreen(), c.getBlue());
             RenderGlobal.drawSelectionBoundingBox(axis);
             GL11.glEnable(3553);
             GL11.glEnable(2929);
