+++ src/main/java/keystrokesmod/utility/RotationUtils.java
@@ -64,7 +64,7 @@ public class RotationUtils {
         final float cos = MathHelper.cos(n2 - 3.1415927f);
         final float sin = MathHelper.sin(n2 - 3.1415927f);
         final float n4 = -MathHelper.cos(n3);
-        final Vec3 vec3 = new Vec3(sin * n4, MathHelper.sin(n3), (double)(cos * n4));
+        final Vec3 vec3 = new Vec3(sin * n4, MathHelper.sin(n3), cos * n4);
         return BlockUtils.getBlock(blockPos).getCollisionBoundingBox(mc.theWorld, blockPos, BlockUtils.getBlockState(blockPos)).calculateIntercept(getPositionEyes, getPositionEyes.addVector(vec3.xCoord * n, vec3.yCoord * n, vec3.zCoord * n)) != null;
     }
 
@@ -149,7 +149,7 @@ public class RotationUtils {
         final float cos = MathHelper.cos(n4 - 3.1415927f);
         final float sin = MathHelper.sin(n4 - 3.1415927f);
         final float n6 = -MathHelper.cos(n5);
-        final Vec3 vec3 = new Vec3((double)(sin * n6), (double)MathHelper.sin(n5), (double)(cos * n6));
+        final Vec3 vec3 = new Vec3(sin * n6, MathHelper.sin(n5), cos * n6);
         return mc.theWorld.rayTraceBlocks(getPositionEyes, getPositionEyes.addVector(vec3.xCoord * n, vec3.yCoord * n, vec3.zCoord * n), false, false, false);
     }
